visualization_service/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile visualization_service/requirements.in -o visualization_service/requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
click==8.1.7
    # via
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
contourpy==1.3.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
dnspython==2.6.1
    # via email-validator
email-validator==2.1.2
    # via fastapi
fastapi==0.112.2
    # via -r visualization_service/requirements.in
fastapi-cli==0.0.5
    # via fastapi
fonttools==4.54.1
    # via matplotlib
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.2
    # via fastapi
idna==3.8
    # via
    #   anyio
    #   email-validator
    #   httpx
itsdangerous==2.2.0
    # via fastapi
jinja2==3.1.4
    # via fastapi
kiwisolver==1.4.7
    # via matplotlib
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
matplotlib==3.9.2
    # via
    #   -r visualization_service/requirements.in
    #   seaborn
mdurl==0.1.2
    # via markdown-it-py
numpy==2.1.3
    # via
    #   contourpy
    #   matplotlib
    #   pandas
    #   seaborn
orjson==3.10.7
    # via fastapi
packaging==24.1
    # via matplotlib
pandas==2.2.3
    # via
    #   -r visualization_service/requirements.in
    #   seaborn
pillow==11.0.0
    # via matplotlib
pydantic==2.8.2
    # via
    #   fastapi
    #   pydantic-extra-types
    #   pydantic-settings
pydantic-core==2.20.1
    # via pydantic
pydantic-extra-types==2.9.0
    # via fastapi
pydantic-settings==2.4.0
    # via fastapi
pygments==2.18.0
    # via rich
pyparsing==3.2.0
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.9
    # via fastapi
pytz==2024.2
    # via pandas
pyyaml==6.0.2
    # via
    #   fastapi
    #   uvicorn
redis==5.2.0
    # via -r visualization_service/requirements.in
rich==13.8.0
    # via typer
seaborn==0.13.2
    # via -r visualization_service/requirements.in
shellingham==1.5.4
    # via typer
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
starlette==0.38.3
    # via fastapi
typer==0.12.5
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typer
tzdata==2024.2
    # via pandas
ujson==5.10.0
    # via fastapi
uvicorn==0.30.6
    # via
    #   fastapi
    #   fastapi-cli
watchfiles==0.24.0
    # via uvicorn
websockets==13.0.1
    # via uvicorn
